+++ packages/visual-stack/src/experimental/Box/index.js
@@ -17,7 +17,6 @@ const Box = ({
   paddingLeft,
   paddingRight,
   paddingTop,
-  wrap,
   ...restProps
 }) => {
   const classes = classNames(
@@ -41,7 +40,6 @@ const Box = ({
     alignItems && `vs-box-align-items-${alignItems}`,
     justifyContent && `vs-box-justify-content-${justifyContent}`,
     grow && 'vs-box-grow',
-    wrap && 'vs-box-wrap',
     className
   );
   return (
@@ -71,7 +69,6 @@ Box.propTypes = {
   border: PropTypes.bool,
   gap: PropTypes.oneOf([true, 'small', 'large']),
   grow: PropTypes.bool,
-  wrap: PropTypes.bool,
   padding: PropTypes.oneOf([true, 'small', 'large']),
   paddingBottom: PropTypes.oneOf([true, 'small', 'large']),
   paddingLeft: PropTypes.oneOf([true, 'small', 'large']),